+++ pkg/storage/stores/tsdb/index_client_test.go
@@ -280,10 +280,10 @@ func TestIndexClient_SeriesVolume(t *testing.T) {
 
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{foo="bar"}`, Value: "", Volume: 200 * 1024},
-				{Name: `{fizz="buzz", foo="bar"}`, Value: "", Volume: 100 * 1024},
-				{Name: `{fizz="buzz"}`, Value: "", Volume: 100 * 1024},
-				{Name: `{ping="pong"}`, Value: "", Volume: 100 * 1024},
+				{Name: `{foo="bar"}`, Volume: 200 * 1024},
+				{Name: `{fizz="buzz", foo="bar"}`, Volume: 100 * 1024},
+				{Name: `{fizz="buzz"}`, Volume: 100 * 1024},
+				{Name: `{ping="pong"}`, Volume: 100 * 1024},
 			},
 			Limit:   10,
 			From:    from,
@@ -297,7 +297,7 @@ func TestIndexClient_SeriesVolume(t *testing.T) {
 
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{foo="bar"}`, Value: "", Volume: 200 * 1024},
+				{Name: `{foo="bar"}`, Volume: 200 * 1024},
 			},
 			Limit:   1,
 			From:    from,