+++ pkg/storage/stores/tsdb/single_file_index_test.go
@@ -438,8 +438,8 @@ func TestTSDBIndex_SeriesVolume(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{fizz="fizz", foo="bar"}`, Value: "", Volume: (30 + 40) * 1024},
-				{Name: `{fizz="buzz", foo="bar"}`, Value: "", Volume: (10 + 20) * 1024},
+				{Name: `{fizz="fizz", foo="bar"}`, Volume: (30 + 40) * 1024},
+				{Name: `{fizz="buzz", foo="bar"}`, Volume: (10 + 20) * 1024},
 			},
 			Limit:   10,
 			From:    from,
@@ -457,8 +457,8 @@ func TestTSDBIndex_SeriesVolume(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{fizz="fizz", foo="bar"}`, Value: "", Volume: (30 + 40) * 1024},
-				{Name: `{fizz="buzz", foo="bar"}`, Value: "", Volume: (10 + 20) * 1024},
+				{Name: `{fizz="fizz", foo="bar"}`, Volume: (30 + 40) * 1024},
+				{Name: `{fizz="buzz", foo="bar"}`, Volume: (10 + 20) * 1024},
 			},
 			Limit:   10,
 			From:    from,
@@ -486,7 +486,7 @@ func TestTSDBIndex_SeriesVolume(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{foo="bar"}`, Value: "", Volume: (10 + 20 + 30 + 40) * 1024},
+				{Name: `{foo="bar"}`, Volume: (10 + 20 + 30 + 40) * 1024},
 			},
 			Limit:   10,
 			From:    from,
@@ -504,8 +504,8 @@ func TestTSDBIndex_SeriesVolume(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{fizz="fizz", foo="bar"}`, Value: "", Volume: (30 + 40) * 1024},
-				{Name: `{fizz="buzz", foo="bar"}`, Value: "", Volume: (10 + 20) * 1024},
+				{Name: `{fizz="fizz", foo="bar"}`, Volume: (30 + 40) * 1024},
+				{Name: `{fizz="buzz", foo="bar"}`, Volume: (10 + 20) * 1024},
 			},
 			Limit:   10,
 			From:    from,
@@ -523,7 +523,7 @@ func TestTSDBIndex_SeriesVolume(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, &logproto.VolumeResponse{
 			Volumes: []logproto.Volume{
-				{Name: `{fizz="fizz", foo="bar"}`, Value: "", Volume: (30 + 40) * 1024},
+				{Name: `{fizz="fizz", foo="bar"}`, Volume: (30 + 40) * 1024},
 				{Name: `{fizz="buzz", foo="bar"}`, Volume: (10 + 20) * 1024},
 			},
 			Limit:   10,